linkedln-scraper/requirements.txt
selenium 
pandas 
python-dotenv 
webdriver-manager